+++ types/base-types.d.ts
@@ -1,1 +1,7 @@
-declare type TypeRef = string | Function;
+declare type TypeRef =
+    string | Function | ArrayBufferConstructor | ArrayConstructor |
+    BooleanConstructor | StringConstructor |
+    NumberConstructor | MapConstructor |
+    SetConstructor | WeakMapConstructor |
+    WeakSetConstructor | PromiseConstructorLike
+    ;